note

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: kurosawaitsuki <kurosawaitsuki@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/06 20:19:32 by kurosawaits       #+#    #+#             */
-/*   Updated: 2022/11/06 20:19:33 by kurosawaits      ###   ########.fr       */
+/*   Updated: 2022/11/23 12:21:53 by kurosawaits      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	if (s && f)
@@ -26,3 +26,5 @@ void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 		}	
 	}
 }
+// 関数ポインタが指す関数の第一引数がunsigned int型なので変換する.
+// size_t型よりもunsigned int型の方が最大値大きい.
